--- Assets/Scripts/Terrain/PathGraph/CostApplicator.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using UnityEngine;
-using Random = System.Random;
-using Vector2 = System.Numerics.Vector2;
-
-namespace Terrain.PathGraph
-{
-    public class CostApplicator
-    {
-        // public void 
-    }
-}
