+++ Assets/Scripts/Terrain/PathGraph/Graphs/Path.cs
@@ -1,10 +1,8 @@
-﻿using System;
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using UnityEngine;
 
-namespace Terrain.PathGraph
+namespace Terrain.PathGraph.Graphs
 {
     /**
      * Linear directed graph
