+++ Assets/Scripts/Terrain/PathGraph/Graphs/Points/CostApplicator.cs
@@ -1,0 +1,7 @@
+﻿namespace Terrain.PathGraph.Graphs.Points
+{
+    public class CostApplicator
+    {
+        // public void 
+    }
+}
