+++ Assets/Scripts/Terrain/PathGraph/PathFinder.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
+using Terrain.PathGraph.Graphs;
 using Unity.Mathematics;
 using UnityEngine;
 using Random = System.Random;
